+++ packages/components/src/components/link-pure/link-pure-utils.ts
@@ -10,6 +10,7 @@ import type { LinkAriaAttribute } from '../../utils';
 
 export type LinkPureIcon = LinkButtonIconName;
 export type LinkPureAriaAttribute = LinkAriaAttribute;
+/** @deprecated */
 export type LinkPureAlignLabelDeprecated = AlignLabelDeprecated;
 export type LinkPureAlignLabel = AlignLabel;
 export type LinkPureWeight = TextWeight;