+++ packages/components/src/components/switch/switch-utils.ts
@@ -1,8 +1,10 @@
 import type { AlignLabel, AlignLabelDeprecated, AriaAttributes } from '../../types';
 import { getButtonBaseAriaAttributes } from '../../utils';
 
+/** @deprecated */
 export type SwitchAlignLabelDeprecated = AlignLabelDeprecated;
 export type SwitchAlignLabel = AlignLabel;
+
 export type SwitchUpdateEvent = { checked: boolean };
 
 export const getSwitchButtonAriaAttributes = (